Data.Yaml.Builder

Description

drastically, or be entirely removed, in a future release.

Documentation

newtype YamlBuilder

Constructors

YamlBuilder

Fields

unYamlBuilder :: [Event] -> [Event]

class ToYaml a where

Methods

toYaml :: a -> YamlBuilder

mapping :: [(Text, YamlBuilder)] -> YamlBuilder

Since: 0.8.7

namedMapping :: Text -> [(Text, YamlBuilder)] -> YamlBuilder

Since: 0.10.3.0

maybeNamedMapping :: Maybe Text -> [(Text, YamlBuilder)] -> YamlBuilder

Since: 0.10.3.0

mappingComplex :: [(YamlBuilder, YamlBuilder)] -> YamlBuilder

Since: 0.11.2.0

namedMappingComplex :: Text -> [(YamlBuilder, YamlBuilder)] -> YamlBuilder

Since: 0.11.2.0

maybeNamedMappingComplex :: Maybe Text -> [(YamlBuilder, YamlBuilder)] -> YamlBuilder

Since: 0.11.2.0

array :: [YamlBuilder] -> YamlBuilder

Since: 0.8.7

namedArray :: Text -> [YamlBuilder] -> YamlBuilder

Since: 0.10.3.0

maybeNamedArray :: Maybe Text -> [YamlBuilder] -> YamlBuilder

Since: 0.10.3.0

string :: Text -> YamlBuilder

Since: 0.8.7

namedString :: Text -> Text -> YamlBuilder

Since: 0.10.3.0

maybeNamedString :: Maybe Text -> Text -> YamlBuilder

Since: 0.10.3.0

bool :: Bool -> YamlBuilder

Since: 0.8.13

namedBool :: Text -> Bool -> YamlBuilder

Since: 0.10.3.0

maybeNamedBool :: Maybe Text -> Bool -> YamlBuilder

Since: 0.10.3.0

null :: YamlBuilder

Since: 0.8.13

namedNull :: Text -> YamlBuilder

Since: 0.10.3.0

maybeNamedNull :: Maybe Text -> YamlBuilder

Since: 0.10.3.0

scientific :: Scientific -> YamlBuilder

Since: 0.8.13

namedScientific :: Text -> Scientific -> YamlBuilder

Since: 0.10.3.0

maybeNamedScientific :: Maybe Text -> Scientific -> YamlBuilder

Since: 0.10.3.0

alias :: Text -> YamlBuilder

Since: 0.10.3.0

number :: Scientific -> YamlBuilder

Deprecated: Use scientific

Since: 0.8.13

toByteString :: ToYaml a => a -> ByteString

Since: 0.8.7

toByteStringWith :: ToYaml a => FormatOptions -> a -> ByteString

Since: 0.10.2.0

writeYamlFile :: ToYaml a => FilePath -> a -> IO ()

writeYamlFileWith :: ToYaml a => FormatOptions -> FilePath -> a -> IO ()

Since: 0.10.2.0

(.=) :: ToYaml a => Text -> a -> (Text, YamlBuilder)

data FormatOptions

Contains options relating to the formatting (indendation, width) of the YAML output.

Since: libyaml-0.10.2.0

setWidth :: Maybe Int -> FormatOptions -> FormatOptions

Set the maximum number of columns in the YAML output, or Nothing for infinite. By default, the limit is 80 characters.

Since: libyaml-0.10.2.0