Data.Yaml.Include

Documentation

decodeFile :: FromJSON a => FilePath -> IO (Maybe a)

Like decodeFile but with support for relative and absolute includes.

The syntax for includes follows the form:

somekey: !include ./somefile.yaml

decodeFileEither :: FromJSON a => FilePath -> IO (Either ParseException a)

Like decodeFileEither but with support for relative and absolute includes.

The syntax for includes follows the form:

somekey: !include ./somefile.yaml

decodeFileWithWarnings :: FromJSON a => FilePath -> IO (Either ParseException ([Warning], a))

A version of decodeFileEither that returns warnings along with the parse result.

Since: 0.10.0