Data.Yaml

Description

Provides a high-level interface for processing YAML files.

This module reuses most of the infrastructure from the aeson package. This means that you can use all of the existing tools for JSON processing for processing YAML files. As a result, much of the documentation below mentions JSON; do not let that confuse you, it's intentional.

For the most part, YAML content translates directly into JSON, and therefore there is very little data loss. If you need to deal with YAML more directly (e.g., directly deal with aliases), you should use the Text.Libyaml module instead.

For documentation on the aeson types, functions, classes, and operators, please see the Data.Aeson module of the aeson package.

Look in the examples directory of the source repository for some initial pointers on how to use this library.

Encoding

encode :: ToJSON a => a -> ByteString

Encode a value into its YAML representation.

encodeWith :: ToJSON a => EncodeOptions -> a -> ByteString

Encode a value into its YAML representation with custom styling.

Since: 0.10.2.0

encodeFile :: ToJSON a => FilePath -> a -> IO ()

Encode a value into its YAML representation and save to the given file.

encodeFileWith :: ToJSON a => EncodeOptions -> FilePath -> a -> IO ()

Encode a value into its YAML representation with custom styling and save to the given file.

Since: 0.10.2.0

Decoding

decodeEither' :: FromJSON a => ByteString -> Either ParseException a

More helpful version of decodeEither which returns the YamlException.

Since: 0.8.3

decodeFileEither :: FromJSON a => FilePath -> IO (Either ParseException a)

A version of decodeFile which should not throw runtime exceptions.

Since: 0.8.4

decodeFileWithWarnings :: FromJSON a => FilePath -> IO (Either ParseException ([Warning], a))

A version of decodeFileEither that returns warnings along with the parse result.

Since: 0.10.0

decodeThrow :: (MonadThrow m, FromJSON a) => ByteString -> m a

A version of decodeEither' lifted to MonadThrow

Since: 0.8.31

decodeFileThrow :: (MonadIO m, FromJSON a) => FilePath -> m a

A version of decodeFileEither lifted to MonadIO

Since: 0.8.31

Decoding multiple documents

For situations where we need to be able to parse multiple documents separated by `---` in a YAML stream, these functions decode a list of values rather than a single value.

decodeAllEither' :: FromJSON a => ByteString -> Either ParseException [a]

Like decodeEither', but decode multiple documents.

Since: 0.11.5.0

decodeAllFileEither :: FromJSON a => FilePath -> IO (Either ParseException [a])

Like decodeFileEither, but decode multiple documents.

Since: 0.11.5.0

decodeAllFileWithWarnings :: FromJSON a => FilePath -> IO (Either ParseException ([Warning], [a]))

Like decodeFileWithWarnings, but decode multiple documents.

Since: 0.11.5.0

decodeAllThrow :: (MonadThrow m, FromJSON a) => ByteString -> m [a]

Like decodeThrow, but decode multiple documents.

Since: 0.11.5.0

decodeAllFileThrow :: (MonadIO m, FromJSON a) => FilePath -> m [a]

Like decodeFileThrow, but decode multiple documents.

Since: 0.11.5.0

More control over decoding

decodeHelper :: FromJSON a => ConduitM () Event Parse () -> IO (Either ParseException ([Warning], Either String a))

Types

data Value

A JSON value represented as a Haskell value.

Constructors

Object !Object

Array !Array

String !Text

Number !Scientific

Bool !Bool

Null

data Parser a

A JSON parser. N.B. This might not fit your usual understanding of "parser". Instead you might like to think of Parser as a "parse result", i.e. a parser to which the input has already been applied.

type Object = KeyMap Value

A JSON "object" (key/value map).

type Array = Vector Value

A JSON "array" (sequence).

data ParseException

Constructors

NonScalarKey

UnknownAlias

Fields

_anchorName :: AnchorName

UnexpectedEvent

Fields

_received :: Maybe Event

_expected :: Maybe Event

InvalidYaml (Maybe YamlException)

MultipleDocuments

AesonException String

OtherParseException SomeException

NonStringKey JSONPath

NonStringKeyAlias AnchorName Value

CyclicIncludes

LoadSettingsException FilePath ParseException

prettyPrintParseException :: ParseException -> String

Alternative to show to display a ParseException on the screen. Instead of displaying the data constructors applied to their arguments, a more textual output is returned. For example, instead of printing:

InvalidYaml (Just (YamlParseException {yamlProblem = "did not find expected ',' or '}'", yamlContext = "while parsing a flow mapping", yamlProblemMark = YamlMark {yamlIndex = 42, yamlLine = 2, yamlColumn = 12}})))

It looks more pleasant to print:

YAML parse exception at line 2, column 12,
while parsing a flow mapping:
did not find expected ',' or '}'

Since 0.8.11

data YamlException

Constructors

YamlException String

YamlParseException
problem, context, index, position line, position column

Fields

yamlProblem :: String

yamlContext :: String

yamlProblemMark :: YamlMark

data YamlMark

The pointer position

Constructors

YamlMark

Fields

yamlIndex :: Int

yamlLine :: Int

yamlColumn :: Int

Constructors and accessors

object :: [Pair] -> Value

Create a Value from a list of name/value Pairs. If duplicate keys arise, later keys and their associated values win.

array :: [Value] -> Value

Construct a new Value from a list of Values.

(.=) :: (KeyValue kv, ToJSON v) => Key -> v -> kv infixr 8

(.:) :: FromJSON a => Object -> Key -> Parser a

Retrieve the value associated with the given key of an Object. The result is empty if the key is not present or the value cannot be converted to the desired type.

This accessor is appropriate if the key and value must be present in an object for it to be valid. If the key and value are optional, use .:? instead.

(.:?) :: FromJSON a => Object -> Key -> Parser (Maybe a)

Retrieve the value associated with the given key of an Object. The result is Nothing if the key is not present or if its value is Null, or empty if the value cannot be converted to the desired type.

This accessor is most useful if the key and value can be absent from an object without affecting its validity. If the key and value are mandatory, use .: instead.

(.!=) :: Parser (Maybe a) -> a -> Parser a

Helper for use in combination with .:? to provide default values for optional JSON object fields.

This combinator is most useful if the key and value can be absent from an object without affecting its validity and we know a default value to assign in that case. If the key and value are mandatory, use .: instead.

Example usage:

 v1 <- o .:? "opt_field_with_dfl" .!= "default_val"
 v2 <- o .:  "mandatory_field"
 v3 <- o .:? "opt_field2"

With helpers (since 0.8.23)

withObject :: String -> (Object -> Parser a) -> Value -> Parser a

withObject name f value applies f to the Object when value is an Object and fails otherwise.

Error message example

withObject "MyType" f (String "oops")
-- Error: "parsing MyType failed, expected Object, but encountered String"

withText :: String -> (Text -> Parser a) -> Value -> Parser a

withText name f value applies f to the Text when value is a String and fails otherwise.

Error message example

withText "MyType" f Null
-- Error: "parsing MyType failed, expected String, but encountered Null"

withArray :: String -> (Array -> Parser a) -> Value -> Parser a

withArray expected f value applies f to the Array when value is an Array and fails otherwise.

Error message example

withArray "MyType" f (String "oops")
-- Error: "parsing MyType failed, expected Array, but encountered String"

withScientific :: String -> (Scientific -> Parser a) -> Value -> Parser a

withScientific name f value applies f to the Scientific number when value is a Number and fails using typeMismatch otherwise.

Warning: If you are converting from a scientific to an unbounded type such as Integer you may want to add a restriction on the size of the exponent (see withBoundedScientific) to prevent malicious input from filling up the memory of the target system.

Error message example

withScientific "MyType" f (String "oops")
-- Error: "parsing MyType failed, expected Number, but encountered String"

withBool :: String -> (Bool -> Parser a) -> Value -> Parser a

withBool expected f value applies f to the Value when value is a Boolean and fails otherwise.

Error message example

withBool "MyType" f (String "oops")
-- Error: "parsing MyType failed, expected Boolean, but encountered String"

Parsing

parseMonad :: MonadFail m => (a -> Parser b) -> a -> m b

Deprecated: With the MonadFail split, this function is going to be removed in the future. Please migrate to parseEither.

parseEither :: (a -> Parser b) -> a -> Either String b

Run a Parser with an Either result type. If the parse fails, the Left payload will contain an error message.

parseMaybe :: (a -> Parser b) -> a -> Maybe b

Run a Parser with a Maybe result type.

Classes

class ToJSON a where

A type that can be converted to JSON.

Instances in general must specify toJSON and should (but don't need to) specify toEncoding.

An example type and instance:

-- Allow ourselves to write Text literals.
{-# LANGUAGE OverloadedStrings #-}

data Coord = Coord { x :: Double, y :: Double }

instance ToJSON Coord where
  toJSON (Coord x y) = object ["x" .= x, "y" .= y]

  toEncoding (Coord x y) = pairs ("x" .= x <> "y" .= y)

Instead of manually writing your ToJSON instance, there are two options to do it automatically:

Data.Aeson.TH provides Template Haskell functions which will derive an instance at compile time. The generated instance is optimized for your type so it will probably be more efficient than the following option.

The compiler can provide a default generic implementation for toJSON.

To use the second, simply add a deriving Generic clause to your datatype and declare a ToJSON instance. If you require nothing other than defaultOptions, it is sufficient to write (and this is the only alternative where the default toJSON implementation is sufficient):

{-# LANGUAGE DeriveGeneric #-}

import GHC.Generics

data Coord = Coord { x :: Double, y :: Double } deriving Generic

instance ToJSON Coord where
    toEncoding = genericToEncoding defaultOptions

or more conveniently using the DerivingVia extension

deriving via Generically Coord instance ToJSON Coord

If on the other hand you wish to customize the generic decoding, you have to implement both methods:

customOptions = defaultOptions
                { fieldLabelModifier = map toUpper
                }

instance ToJSON Coord where
    toJSON     = genericToJSON customOptions
    toEncoding = genericToEncoding customOptions

Previous versions of this library only had the toJSON method. Adding toEncoding had two reasons:

toEncoding is more efficient for the common case that the output of toJSON is directly serialized to a ByteString. Further, expressing either method in terms of the other would be non-optimal.

The choice of defaults allows a smooth transition for existing users: Existing instances that do not define toEncoding still compile and have the correct semantics. This is ensured by making the default implementation of toEncoding use toJSON. This produces correct results, but since it performs an intermediate conversion to a Value, it will be less efficient than directly emitting an Encoding. (this also means that specifying nothing more than instance ToJSON Coord would be sufficient as a generically decoding instance, but there probably exists no good reason to not specify toEncoding in new instances.)

Minimal complete definition

Nothing

Methods

toJSON :: a -> Value

Convert a Haskell value to a JSON-friendly intermediate type.

toEncoding :: a -> Encoding

Encode a Haskell value as JSON.

The default implementation of this method creates an intermediate Value using toJSON. This provides source-level compatibility for people upgrading from older versions of this library, but obviously offers no performance advantage.

To benefit from direct encoding, you must provide an implementation for this method. The easiest way to do so is by having your types implement Generic using the DeriveGeneric extension, and then have GHC generate a method body as follows.

instance ToJSON Coord where
    toEncoding = genericToEncoding defaultOptions

toJSONList :: [a] -> Value

toEncodingList :: [a] -> Encoding

class FromJSON a where

A type that can be converted from JSON, with the possibility of failure.

In many cases, you can get the compiler to generate parsing code for you (see below). To begin, let's cover writing an instance by hand.

There are various reasons a conversion could fail. For example, an Object could be missing a required key, an Array could be of the wrong size, or a value could be of an incompatible type.

The basic ways to signal a failed conversion are as follows:

fail yields a custom error message: it is the recommended way of reporting a failure;

empty (or mzero) is uninformative: use it when the error is meant to be caught by some (<|>);

typeMismatch can be used to report a failure when the encountered value is not of the expected JSON type; unexpected is an appropriate alternative when more than one type may be expected, or to keep the expected type implicit.

prependFailure (or modifyFailure) add more information to a parser's error messages.

An example type and instance using typeMismatch and prependFailure:

-- Allow ourselves to write Text literals.
{-# LANGUAGE OverloadedStrings #-}

data Coord = Coord { x :: Double, y :: Double }

instance FromJSON Coord where
    parseJSON (Object v) = Coord
        <$> v .: "x"
        <*> v .: "y"

    -- We do not expect a non-Object value here.
    -- We could use empty to fail, but typeMismatch
    -- gives a much more informative error message.
    parseJSON invalid    =
        prependFailure "parsing Coord failed, "
            (typeMismatch "Object" invalid)

For this common case of only being concerned with a single type of JSON value, the functions withObject, withScientific, etc. are provided. Their use is to be preferred when possible, since they are more terse. Using withObject, we can rewrite the above instance (assuming the same language extension and data type) as:

instance FromJSON Coord where
    parseJSON = withObject "Coord" $ \v -> Coord
        <$> v .: "x"
        <*> v .: "y"

Instead of manually writing your FromJSON instance, there are two options to do it automatically:

Data.Aeson.TH provides Template Haskell functions which will derive an instance at compile time. The generated instance is optimized for your type so it will probably be more efficient than the following option.

The compiler can provide a default generic implementation for parseJSON.

To use the second, simply add a deriving Generic clause to your datatype and declare a FromJSON instance for your datatype without giving a definition for parseJSON.

For example, the previous example can be simplified to just:

{-# LANGUAGE DeriveGeneric #-}

import GHC.Generics

data Coord = Coord { x :: Double, y :: Double } deriving Generic

instance FromJSON Coord

or using the DerivingVia extension

deriving via Generically Coord instance FromJSON Coord

The default implementation will be equivalent to parseJSON = genericParseJSON defaultOptions; if you need different options, you can customize the generic decoding by defining:

customOptions = defaultOptions
                { fieldLabelModifier = map toUpper
                }

instance FromJSON Coord where
    parseJSON = genericParseJSON customOptions

Minimal complete definition

Nothing

Methods

parseJSON :: Value -> Parser a

parseJSONList :: Value -> Parser [a]

Custom encoding

isSpecialString :: Text -> Bool

Determine whether a string must be quoted in YAML and can't appear as plain text. Useful if you want to use setStringStyle.

Since: 0.10.2.0

data EncodeOptions

Since: 0.10.2.0

defaultEncodeOptions :: EncodeOptions

Since: 0.10.2.0

defaultStringStyle :: StringStyle

Since: 0.11.2.0

setStringStyle :: (Text -> (Tag, Style)) -> EncodeOptions -> EncodeOptions

Set the string style in the encoded YAML. This is a function that decides for each string the type of YAML string to output.

WARNING: You must ensure that special strings (like "yes"/"no"/"null"/"1234") are not encoded with the Plain style, because then they will be decoded as boolean, null or numeric values. You can use isSpecialString to detect them.

By default, strings are encoded as follows:

Any string containing a newline character uses the Literal style

Otherwise, any special string (see isSpecialString) uses SingleQuoted

Otherwise, use Plain

Since: 0.10.2.0

setFormat :: FormatOptions -> EncodeOptions -> EncodeOptions

Set the encoding formatting for the encoded YAML. By default, this is defaultFormatOptions.

Since: 0.10.2.0

data FormatOptions

Contains options relating to the formatting (indendation, width) of the YAML output.

Since: libyaml-0.10.2.0

defaultFormatOptions :: FormatOptions

Since: libyaml-0.10.2.0

setWidth :: Maybe Int -> FormatOptions -> FormatOptions

Set the maximum number of columns in the YAML output, or Nothing for infinite. By default, the limit is 80 characters.

Since: libyaml-0.10.2.0

Deprecated

decode :: FromJSON a => ByteString -> Maybe a

Deprecated: Please use decodeEither or decodeThrow, which provide information on how the decode failed

decodeFile :: FromJSON a => FilePath -> IO (Maybe a)

Deprecated: Please use decodeFileEither, which does not confused type-directed and runtime exceptions.

decodeEither :: FromJSON a => ByteString -> Either String a

Deprecated: Please use decodeEither' or decodeThrow, which provide more useful failures