Data.Yaml.Include

Documentation

decodeFile :: FromJSON a => FilePath -> IO (Maybe a)

Like decodeFile but with support for relative and absolute includes.

The syntax for includes follows the form:

somekey: !include ./somefile.yaml

decodeFileEither :: FromJSON a => FilePath -> IO (Either ParseException a)

Like decodeFileEither but with support for relative and absolute includes.

The syntax for includes follows the form:

somekey: !include ./somefile.yaml